Fjern ubrukt arkiveringskode
