+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/archive/PeriodicVarselArchiver.kt
@@ -26,7 +26,6 @@ class PeriodicVarselArchiver(
         val thresholdDate = nowAtUtc().minusDays(ageThresholdDays.toLong())
 
         try {
-
             val archivedBeskjeder = varselArchivingRepository.archiveOldBeskjeder(thresholdDate)
             archiveMetricsProbe.countEntitiesArchived(EventType.BESKJED_INTERN, archivedBeskjeder)
 
@@ -39,7 +38,7 @@ class PeriodicVarselArchiver(
         } catch (rt: RetriableDatabaseException) {
             log.warn("Fikk en periodisk feil mot databasen ved arkivering av Beskjed. Forsøker igjen senere.", rt)
         } catch (e: Exception) {
-            log.error("Fikk feil mot databasen ved arkivering av beskjed. Stopper prosessering.", e)
+            log.error("Fikk feil mot databasen ved arkivering av beskjed. Forsøker igjen senere.", e)
         }
     }
 }
