+++ src/test/kotlin/no/nav/personbruker/dittnav/eventaggregator/archive/varselArchiveTestQueries.kt
@@ -1,14 +1,25 @@
-package no.nav.personbruker.dittnav.eventaggregator.innboks.archive
+package no.nav.personbruker.dittnav.eventaggregator.archive
 
-import no.nav.personbruker.dittnav.eventaggregator.archive.BrukernotifikasjonArchiveDTO
 import no.nav.personbruker.dittnav.eventaggregator.common.database.getUtcDateTime
 import no.nav.personbruker.dittnav.eventaggregator.common.database.list
 import java.sql.Connection
 import java.sql.ResultSet
 
-fun Connection.deleteAllInnboksArchive() =
-    prepareStatement("""DELETE FROM INNBOKS_ARKIV""")
-        .use {it.execute()}
+fun Connection.getAllArchivedBeskjed(): List<BrukernotifikasjonArchiveDTO> =
+    prepareStatement("""SELECT * FROM beskjed_arkiv""")
+        .use {
+            it.executeQuery().list {
+                toArchiveDto()
+            }
+        }
+
+fun Connection.getAllArchivedOppgave(): List<BrukernotifikasjonArchiveDTO> =
+    prepareStatement("""SELECT * FROM oppgave_arkiv""")
+        .use {
+            it.executeQuery().list {
+                toArchiveDto()
+            }
+        }
 
 fun Connection.getAllArchivedInnboks(): List<BrukernotifikasjonArchiveDTO> =
     prepareStatement("""SELECT * FROM innboks_arkiv""")
@@ -18,7 +29,7 @@ fun Connection.getAllArchivedInnboks(): List<BrukernotifikasjonArchiveDTO> =
             }
         }
 
-private fun ResultSet.toArchiveDto() =
+fun ResultSet.toArchiveDto() =
     BrukernotifikasjonArchiveDTO(
         eventId = getString("eventid"),
         fodselsnummer = getString("fodselsnummer"),
